refactor: split admin ui into smaller modules
